+++ framework/model/src/test/java/is/codion/framework/model/DefaultEntitySearchModelTest.java
@@ -262,10 +262,10 @@ public final class DefaultEntitySearchModelTest {
 		Map<Entity, Entity> updated = new HashMap<>();
 		updated.put(temp, tempUpdated);
 
-		EntityEditModel.events().updated(Employee.TYPE).accept(updated);
+		EntityEditModel.events(Employee.TYPE).updated().accept(updated);
 		assertEquals("Newname", searchModel.selection().entity().get().get(Employee.NAME));
 
-		EntityEditModel.events().deleted(Employee.TYPE).accept(singletonList(temp));
+		EntityEditModel.events(Employee.TYPE).deleted().accept(singletonList(temp));
 		assertTrue(searchModel.selection().empty().is());
 	}